Further implementation

+++ app/src/main/java/com/example/myapplication4/entity/Person.java
@@ -1,11 +1,11 @@
 package com.example.myapplication4.entity;
 
 public abstract class Person {
-    public int id;
+    public long id;
     public String firstName;
     public String lastName;
 
-    public Person(int id, String firstName, String lastName) {
+    public Person(long id, String firstName, String lastName) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
