+++ app/src/main/java/com/example/myapplication4/entity/Student.java
@@ -3,7 +3,7 @@ package com.example.myapplication4.entity;
 public class Student extends Person {
     public int grade;
 
-    public Student(int id, String firstName, String lastName, int grade) {
+    public Student(long id, String firstName, String lastName, int grade) {
         super(id, firstName, lastName);
         this.grade = grade;
     }
