+++ app/src/main/java/com/example/myapplication4/entity/Teacher.java
@@ -3,7 +3,7 @@ package com.example.myapplication4.entity;
 public class Teacher extends Person {
     public String specialty;
 
-    public Teacher(int id, String firstName, String lastName, String specialty) {
+    public Teacher(long id, String firstName, String lastName, String specialty) {
         super(id, firstName, lastName);
         this.specialty = specialty;
     }
